+++ domain/src/main/java/com/codeinglogs/domain/model/movies/movieslist/Movies.kt
@@ -1,7 +1,7 @@
 package com.codeinglogs.domain.model.movies.movieslist
 
 data class Movies(
-    val id: Double?,
+    val id: Long?,
     val title: String?,
     val poster_path: String?,
     val vote_average: Double?,
