+++ remote/src/main/java/com/codeinglogs/remote/model/movies/movieslist/Movies.kt
@@ -3,7 +3,7 @@ package com.codeinglogs.remote.model.movies.movieslist
 import com.codeinglogs.data.model.movies.movieslist.Movies as DataMovies
 
 data class Movies(
-    val id: Double?,
+    val id: Long?,
     val original_language: String?,
     val original_name: String?,
     val original_title: String?,
@@ -14,7 +14,7 @@ data class Movies(
 )
 
 fun Movies.toDataMovies() = DataMovies(
-    id = id?:0.0,
+    id = id?:0L,
     title = original_name?:original_title,
     poster_path = poster_path,
     vote_average=vote_average
